+++ packages/@expo/cli/e2e/__tests__/export/url-polyfill.test.ts
@@ -1,0 +1,66 @@
+/* eslint-env jest */
+import JsonFile from '@expo/json-file';
+import execa from 'execa';
+import klawSync from 'klaw-sync';
+import path from 'path';
+
+import { runExportSideEffects } from './export-side-effects';
+import { bin, getRouterE2ERoot } from '../utils';
+
+runExportSideEffects();
+
+describe('exports with url-polyfill', () => {
+  const projectRoot = getRouterE2ERoot();
+  const outputName = 'dist-url-polyfill';
+  const outputDir = path.join(projectRoot, outputName);
+
+  beforeAll(
+    async () => {
+      await execa('node', [bin, 'export', '-p', 'ios', '--output-dir', outputName], {
+        cwd: projectRoot,
+        env: {
+          NODE_ENV: 'production',
+          EXPO_USE_STATIC: '1',
+          E2E_ROUTER_SRC: 'url-polyfill',
+          E2E_ROUTER_ASYNC: 'development',
+        },
+      });
+    },
+    // Could take 45s depending on how fast the bundler resolves
+    560 * 1000
+  );
+
+  it('has expected files', async () => {
+    // List output files with sizes for snapshotting.
+    // This is to make sure that any changes to the output are intentional.
+    // Posix path formatting is used to make paths the same across OSes.
+    const files = klawSync(outputDir)
+      .map((entry) => {
+        if (entry.path.includes('node_modules') || !entry.stats.isFile()) {
+          return null;
+        }
+        return path.posix.relative(outputDir, entry.path);
+      })
+      .filter(Boolean);
+
+    const metadata = await JsonFile.readAsync(path.resolve(outputDir, 'metadata.json'));
+
+    expect(metadata).toEqual({
+      bundler: 'metro',
+      fileMetadata: {
+        ios: {
+          assets: expect.anything(),
+          bundle: expect.stringMatching(/bundles\/ios-.*\.js/),
+        },
+      },
+      version: 0,
+    });
+
+    // The wrapper should not be included as a route.
+    expect(files).not.toContain('+html.html');
+    expect(files).not.toContain('index.html');
+
+    const iosBundle = files.find((v) => v?.startsWith('bundles/ios'));
+    expect(iosBundle).toBeDefined();
+  });
+});
